requirements.txt
blinker==1.7.0
Click==8.1.7
Faker==24.9.0
Flask==1.1.2
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
Flask-Uploads==0.2.1
ipykernel==6.29.3
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector-python==8.3.0
mysqlclient==2.2.4
passlib==1.7.1
tornado==6.1
Werkzeug==0.16
wheel==0.41.2
WTForms==2.2.1
